+++ packages/frontend/src/pages/zk-catalog/v2/components/TrustedSetupCell.tsx
@@ -1,4 +1,5 @@
 import type { TrustedSetup, ZkCatalogTag } from '@l2beat/config'
+import { assert } from '@l2beat/shared-pure'
 import {
   Tooltip,
   TooltipContent,
@@ -21,48 +22,10 @@ interface Props {
 export function TrustedSetupCell({ trustedSetups, projectsUsedIn }: Props) {
   const proofSystem = trustedSetups[0]?.proofSystem
   if (trustedSetups.length === 0 || !proofSystem) return null
-  const worstRisk = pickWorstRisk(trustedSetups)
 
   return (
     <div className="flex flex-col items-start gap-2">
-      <Tooltip>
-        <TooltipTrigger className="flex items-center gap-2">
-          {worstRisk === 'N/A' ? (
-            <span className="text-2xl leading-none">🤩</span>
-          ) : (
-            <TrustedSetupRiskDot risk={worstRisk} size="md" />
-          )}
-          <TechStackTag tag={proofSystem} withoutTooltip />
-        </TooltipTrigger>
-        <TooltipContent>
-          <div className="mb-3 text-paragraph-14">
-            Trusted setups for{' '}
-            <TechStackTag
-              tag={proofSystem}
-              className="inline-block"
-              withoutTooltip
-            />
-          </div>
-          {trustedSetups.map((trustedSetup) => {
-            return (
-              <div key={trustedSetup.id} className="flex gap-2">
-                {trustedSetup.risk === 'N/A' ? (
-                  <div className="mt-px text-lg leading-none">🤩</div>
-                ) : (
-                  <TrustedSetupRiskDot
-                    risk={trustedSetup.risk}
-                    size="sm"
-                    className="shrink-0"
-                  />
-                )}
-                <span className="text-xs leading-normal">
-                  {trustedSetup.shortDescription}
-                </span>
-              </div>
-            )
-          })}
-        </TooltipContent>
-      </Tooltip>
+      <TrustedSetupCellTooltip trustedSetups={trustedSetups} />
       {projectsUsedIn && (
         <div className="flex items-center gap-1.5">
           <p className="font-medium text-label-value-12 text-secondary">
@@ -75,6 +38,59 @@ export function TrustedSetupCell({ trustedSetups, projectsUsedIn }: Props) {
         </div>
       )}
     </div>
+  )
+}
+
+export function TrustedSetupCellTooltip({
+  trustedSetups,
+}: {
+  trustedSetups: (TrustedSetup & {
+    proofSystem: ZkCatalogTag
+  })[]
+}) {
+  const proofSystem = trustedSetups[0]?.proofSystem
+  assert(proofSystem, 'proofSystem is required')
+  const worstRisk = pickWorstRisk(trustedSetups)
+
+  return (
+    <Tooltip>
+      <TooltipTrigger className="flex items-center gap-2">
+        {worstRisk === 'N/A' ? (
+          <span className="text-2xl leading-none">🤩</span>
+        ) : (
+          <TrustedSetupRiskDot risk={worstRisk} size="md" />
+        )}
+        <TechStackTag tag={proofSystem} withoutTooltip />
+      </TooltipTrigger>
+      <TooltipContent>
+        <div className="mb-3 text-paragraph-14">
+          Trusted setups for{' '}
+          <TechStackTag
+            tag={proofSystem}
+            className="inline-block"
+            withoutTooltip
+          />
+        </div>
+        {trustedSetups.map((trustedSetup) => {
+          return (
+            <div key={trustedSetup.id} className="flex gap-2">
+              {trustedSetup.risk === 'N/A' ? (
+                <div className="mt-px text-lg leading-none">🤩</div>
+              ) : (
+                <TrustedSetupRiskDot
+                  risk={trustedSetup.risk}
+                  size="sm"
+                  className="shrink-0"
+                />
+              )}
+              <span className="text-xs leading-normal">
+                {trustedSetup.shortDescription}
+              </span>
+            </div>
+          )
+        })}
+      </TooltipContent>
+    </Tooltip>
   )
 }
