+++ packages/frontend/src/server/features/zk-catalog/project/getZkCatalogProjectEntry.ts
@@ -2,17 +2,23 @@ import type { Project } from '@l2beat/config'
 import type { UnixTime } from '@l2beat/shared-pure'
 import type { ProjectLink } from '~/components/projects/links/types'
 import type { ProjectDetailsSection } from '~/components/projects/sections/types'
+import { ps } from '~/server/projects'
 import type { SsrHelpers } from '~/trpc/server'
 import { getProjectLinks } from '~/utils/project/getProjectLinks'
 import {
   getUnderReviewStatus,
   type UnderReviewStatus,
 } from '~/utils/project/underReview'
-import { getProjectIcon } from '../utils/getProjectIcon'
+import { getProjectIcon } from '../../utils/getProjectIcon'
+import {
+  getTrustedSetupsWithVerifiersAndAttesters,
+  type TrustedSetupsByProofSystem,
+} from '../getTrustedSetupsWithVerifiersAndAttesters'
 
 export interface ProjectZkCatalogEntry {
   name: string
   shortName: string | undefined
+  creator?: string
   slug: string
   icon: string
   archivedAt: UnixTime | undefined
@@ -23,6 +29,7 @@ export interface ProjectZkCatalogEntry {
     emergencyWarning?: string
     description?: string
     links: ProjectLink[]
+    trustedSetupsByProofSystem: TrustedSetupsByProofSystem
   }
   sections: ProjectDetailsSection[]
 }
@@ -34,16 +41,27 @@ export async function getZkCatalogProjectEntry(
   >,
   helpers: SsrHelpers,
 ): Promise<ProjectZkCatalogEntry> {
+  const allProjects = await ps.getProjects({
+    optional: ['daBridge', 'isBridge', 'isScaling', 'isDaLayer'],
+  })
+
+  const trustedSetupsByProofSystem = getTrustedSetupsWithVerifiersAndAttesters(
+    project,
+    allProjects,
+  )
+
   const header: ProjectZkCatalogEntry['header'] = {
     description: project.display.description,
     warning: project.statuses.yellowWarning,
     redWarning: project.statuses.redWarning,
     emergencyWarning: project.statuses.emergencyWarning,
     links: getProjectLinks(project.display.links),
+    trustedSetupsByProofSystem,
   }
 
   const common = {
     name: project.name,
+    creator: project.zkCatalogInfo.creator,
     shortName: project.shortName,
     slug: project.slug,
     icon: getProjectIcon(project.slug),
@@ -56,8 +74,6 @@ export async function getZkCatalogProjectEntry(
   }
 
   const sections: ProjectDetailsSection[] = []
-
-  await Promise.resolve()
 
   return { ...common, sections }
 }
